Move `ImageCommentsViewAdapter` to production
